Move settings and configuration to hare.conf module (#24)
