+++ hare/conf/asgi.py
@@ -33,6 +33,6 @@ import os
 
 from django.core.asgi import get_asgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "hare.core.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "hare.conf.settings")
 
 application = get_asgi_application()
