+++ mypy.ini
@@ -2,7 +2,7 @@
 plugins = mypy_django_plugin.main
 
 [mypy.plugins.django-stubs]
-django_settings_module = "hare.core.settings"
+django_settings_module = "hare.conf.settings"
 
 [mypy-hare.core.migrations.*]
 ignore_errors = True
